\name{fpclusters}
\alias{fpclusters}
%- Also NEED an `\alias' for EACH other topic documented here.
\title{Extracting clusters from fixed point cluster objects}
\description{
  \code{fpclusters} is a generic function which extracts the
  representative fixed point clusters (FPCs)
  from FPC objects generated by \code{\link{fixmahal}} and
  \code{\link{fixreg}}. For documentation and examples see
  \code{\link{fixmahal}} and  \code{\link{fixreg}}.
}
\usage{
fpclusters(object, ...)
}
%- maybe also `usage' for other objects documented here.
\arguments{
  \item{object}{object of class \code{rfpc} or \code{mfpc}.}
  \item{...}{further arguments depending on the method.}
}

\value{
  a list of logical or numerical vectors indicating or giving the
  weights of the cluster memberships.
}

\author{Christian Hennig
  \email{c.hennig@ucl.ac.uk}
  \url{http://www.homepages.ucl.ac.uk/~ucakche/}
}

\seealso{\code{\link{fixmahal}}, \code{\link{fixreg}}}

\keyword{cluster}% at least one, from doc/KEYWORDS

